+++ api/golang/core/kurtosis_core_rpc_api_bindings/api_container_service_grpc.pb.go
@@ -32,9 +32,15 @@ type ApiContainerServiceClient interface {
 	// Executes an executable module on the user's behalf
 	ExecuteModule(ctx context.Context, in *ExecuteModuleArgs, opts ...grpc.CallOption) (*ExecuteModuleResponse, error)
 	// Executes a startosis script on the user's behalf
+	// Will soon be deprecated in favour of its streamed version ExecuteKurtosisScript
 	ExecuteStartosisScript(ctx context.Context, in *ExecuteStartosisScriptArgs, opts ...grpc.CallOption) (*ExecuteStartosisResponse, error)
+	// Executes a startosis script on the user's behalf
+	ExecuteKurtosisScript(ctx context.Context, in *ExecuteStartosisScriptArgs, opts ...grpc.CallOption) (ApiContainerService_ExecuteKurtosisScriptClient, error)
 	// Executes a startosis module on the user's behalf
+	// Will soon be deprecated in favour of its streamed version ExecuteKurtosisModule
 	ExecuteStartosisModule(ctx context.Context, in *ExecuteStartosisModuleArgs, opts ...grpc.CallOption) (*ExecuteStartosisResponse, error)
+	// Executes a startosis script on the user's behalf
+	ExecuteKurtosisModule(ctx context.Context, in *ExecuteStartosisModuleArgs, opts ...grpc.CallOption) (ApiContainerService_ExecuteKurtosisModuleClient, error)
 	// Start services by creating containers for them
 	StartServices(ctx context.Context, in *StartServicesArgs, opts ...grpc.CallOption) (*StartServicesResponse, error)
 	// Returns the IDs of the current services in the enclave
@@ -123,6 +129,38 @@ func (c *apiContainerServiceClient) ExecuteStartosisScript(ctx context.Context, 
 	return out, nil
 }
 
+func (c *apiContainerServiceClient) ExecuteKurtosisScript(ctx context.Context, in *ExecuteStartosisScriptArgs, opts ...grpc.CallOption) (ApiContainerService_ExecuteKurtosisScriptClient, error) {
+	stream, err := c.cc.NewStream(ctx, &ApiContainerService_ServiceDesc.Streams[0], "/api_container_api.ApiContainerService/ExecuteKurtosisScript", opts...)
+	if err != nil {
+		return nil, err
+	}
+	x := &apiContainerServiceExecuteKurtosisScriptClient{stream}
+	if err := x.ClientStream.SendMsg(in); err != nil {
+		return nil, err
+	}
+	if err := x.ClientStream.CloseSend(); err != nil {
+		return nil, err
+	}
+	return x, nil
+}
+
+type ApiContainerService_ExecuteKurtosisScriptClient interface {
+	Recv() (*KurtosisExecutionResponseLine, error)
+	grpc.ClientStream
+}
+
+type apiContainerServiceExecuteKurtosisScriptClient struct {
+	grpc.ClientStream
+}
+
+func (x *apiContainerServiceExecuteKurtosisScriptClient) Recv() (*KurtosisExecutionResponseLine, error) {
+	m := new(KurtosisExecutionResponseLine)
+	if err := x.ClientStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
+}
+
 func (c *apiContainerServiceClient) ExecuteStartosisModule(ctx context.Context, in *ExecuteStartosisModuleArgs, opts ...grpc.CallOption) (*ExecuteStartosisResponse, error) {
 	out := new(ExecuteStartosisResponse)
 	err := c.cc.Invoke(ctx, "/api_container_api.ApiContainerService/ExecuteStartosisModule", in, out, opts...)
@@ -130,6 +168,38 @@ func (c *apiContainerServiceClient) ExecuteStartosisModule(ctx context.Context, 
 		return nil, err
 	}
 	return out, nil
+}
+
+func (c *apiContainerServiceClient) ExecuteKurtosisModule(ctx context.Context, in *ExecuteStartosisModuleArgs, opts ...grpc.CallOption) (ApiContainerService_ExecuteKurtosisModuleClient, error) {
+	stream, err := c.cc.NewStream(ctx, &ApiContainerService_ServiceDesc.Streams[1], "/api_container_api.ApiContainerService/ExecuteKurtosisModule", opts...)
+	if err != nil {
+		return nil, err
+	}
+	x := &apiContainerServiceExecuteKurtosisModuleClient{stream}
+	if err := x.ClientStream.SendMsg(in); err != nil {
+		return nil, err
+	}
+	if err := x.ClientStream.CloseSend(); err != nil {
+		return nil, err
+	}
+	return x, nil
+}
+
+type ApiContainerService_ExecuteKurtosisModuleClient interface {
+	Recv() (*KurtosisExecutionResponseLine, error)
+	grpc.ClientStream
+}
+
+type apiContainerServiceExecuteKurtosisModuleClient struct {
+	grpc.ClientStream
+}
+
+func (x *apiContainerServiceExecuteKurtosisModuleClient) Recv() (*KurtosisExecutionResponseLine, error) {
+	m := new(KurtosisExecutionResponseLine)
+	if err := x.ClientStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
 }
 
 func (c *apiContainerServiceClient) StartServices(ctx context.Context, in *StartServicesArgs, opts ...grpc.CallOption) (*StartServicesResponse, error) {
@@ -289,9 +359,15 @@ type ApiContainerServiceServer interface {
 	// Executes an executable module on the user's behalf
 	ExecuteModule(context.Context, *ExecuteModuleArgs) (*ExecuteModuleResponse, error)
 	// Executes a startosis script on the user's behalf
+	// Will soon be deprecated in favour of its streamed version ExecuteKurtosisScript
 	ExecuteStartosisScript(context.Context, *ExecuteStartosisScriptArgs) (*ExecuteStartosisResponse, error)
+	// Executes a startosis script on the user's behalf
+	ExecuteKurtosisScript(*ExecuteStartosisScriptArgs, ApiContainerService_ExecuteKurtosisScriptServer) error
 	// Executes a startosis module on the user's behalf
+	// Will soon be deprecated in favour of its streamed version ExecuteKurtosisModule
 	ExecuteStartosisModule(context.Context, *ExecuteStartosisModuleArgs) (*ExecuteStartosisResponse, error)
+	// Executes a startosis script on the user's behalf
+	ExecuteKurtosisModule(*ExecuteStartosisModuleArgs, ApiContainerService_ExecuteKurtosisModuleServer) error
 	// Start services by creating containers for them
 	StartServices(context.Context, *StartServicesArgs) (*StartServicesResponse, error)
 	// Returns the IDs of the current services in the enclave
@@ -346,8 +422,14 @@ func (UnimplementedApiContainerServiceServer) ExecuteModule(context.Context, *Ex
 func (UnimplementedApiContainerServiceServer) ExecuteStartosisScript(context.Context, *ExecuteStartosisScriptArgs) (*ExecuteStartosisResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ExecuteStartosisScript not implemented")
 }
+func (UnimplementedApiContainerServiceServer) ExecuteKurtosisScript(*ExecuteStartosisScriptArgs, ApiContainerService_ExecuteKurtosisScriptServer) error {
+	return status.Errorf(codes.Unimplemented, "method ExecuteKurtosisScript not implemented")
+}
 func (UnimplementedApiContainerServiceServer) ExecuteStartosisModule(context.Context, *ExecuteStartosisModuleArgs) (*ExecuteStartosisResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ExecuteStartosisModule not implemented")
+}
+func (UnimplementedApiContainerServiceServer) ExecuteKurtosisModule(*ExecuteStartosisModuleArgs, ApiContainerService_ExecuteKurtosisModuleServer) error {
+	return status.Errorf(codes.Unimplemented, "method ExecuteKurtosisModule not implemented")
 }
 func (UnimplementedApiContainerServiceServer) StartServices(context.Context, *StartServicesArgs) (*StartServicesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method StartServices not implemented")
@@ -499,6 +581,27 @@ func _ApiContainerService_ExecuteStartosisScript_Handler(srv interface{}, ctx co
 	return interceptor(ctx, in, info, handler)
 }
 
+func _ApiContainerService_ExecuteKurtosisScript_Handler(srv interface{}, stream grpc.ServerStream) error {
+	m := new(ExecuteStartosisScriptArgs)
+	if err := stream.RecvMsg(m); err != nil {
+		return err
+	}
+	return srv.(ApiContainerServiceServer).ExecuteKurtosisScript(m, &apiContainerServiceExecuteKurtosisScriptServer{stream})
+}
+
+type ApiContainerService_ExecuteKurtosisScriptServer interface {
+	Send(*KurtosisExecutionResponseLine) error
+	grpc.ServerStream
+}
+
+type apiContainerServiceExecuteKurtosisScriptServer struct {
+	grpc.ServerStream
+}
+
+func (x *apiContainerServiceExecuteKurtosisScriptServer) Send(m *KurtosisExecutionResponseLine) error {
+	return x.ServerStream.SendMsg(m)
+}
+
 func _ApiContainerService_ExecuteStartosisModule_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ExecuteStartosisModuleArgs)
 	if err := dec(in); err != nil {
@@ -515,6 +618,27 @@ func _ApiContainerService_ExecuteStartosisModule_Handler(srv interface{}, ctx co
 		return srv.(ApiContainerServiceServer).ExecuteStartosisModule(ctx, req.(*ExecuteStartosisModuleArgs))
 	}
 	return interceptor(ctx, in, info, handler)
+}
+
+func _ApiContainerService_ExecuteKurtosisModule_Handler(srv interface{}, stream grpc.ServerStream) error {
+	m := new(ExecuteStartosisModuleArgs)
+	if err := stream.RecvMsg(m); err != nil {
+		return err
+	}
+	return srv.(ApiContainerServiceServer).ExecuteKurtosisModule(m, &apiContainerServiceExecuteKurtosisModuleServer{stream})
+}
+
+type ApiContainerService_ExecuteKurtosisModuleServer interface {
+	Send(*KurtosisExecutionResponseLine) error
+	grpc.ServerStream
+}
+
+type apiContainerServiceExecuteKurtosisModuleServer struct {
+	grpc.ServerStream
+}
+
+func (x *apiContainerServiceExecuteKurtosisModuleServer) Send(m *KurtosisExecutionResponseLine) error {
+	return x.ServerStream.SendMsg(m)
 }
 
 func _ApiContainerService_StartServices_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
@@ -901,6 +1025,17 @@ var ApiContainerService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _ApiContainerService_RenderTemplatesToFilesArtifact_Handler,
 		},
 	},
-	Streams:  []grpc.StreamDesc{},
+	Streams: []grpc.StreamDesc{
+		{
+			StreamName:    "ExecuteKurtosisScript",
+			Handler:       _ApiContainerService_ExecuteKurtosisScript_Handler,
+			ServerStreams: true,
+		},
+		{
+			StreamName:    "ExecuteKurtosisModule",
+			Handler:       _ApiContainerService_ExecuteKurtosisModule_Handler,
+			ServerStreams: true,
+		},
+	},
 	Metadata: "api_container_service.proto",
 }