+++ cli/cli/commands/exec/exec.go
@@ -41,6 +41,7 @@ const (
 	scriptArgForLogging = "script"
 	moduleArgForLogging = "module"
 
+	githubDomainPrefix = "github.com/"
 	isNewEnclave = true
 )
 
@@ -55,7 +56,7 @@ var StartosisExecCmd = &lowlevel.LowlevelKurtosisCommand{
 	CommandStr:       command_str_consts.StartosisExecCmdStr,
 	ShortDescription: "Execute a Startosis script or module",
 	LongDescription: "Execute a Startosis module or script in an enclave. For a script we expect a path to a " + startosisExtension +
-		" file. For a module we expect path to a directory containing kurtosis.mod. If the enclave-id param is provided, Kurtosis " +
+		" file. For a module we expect path to a directory containing kurtosis.mod or a fully qualified Github repository path containing a module. If the enclave-id param is provided, Kurtosis " +
 		"will exec the script inside this enclave, or create it if it doesn't exist. If no enclave-id param is " +
 		"provided, Kurtosis will create a new enclave with a default name derived from the script or module name.",
 	Flags: []*flags.FlagConfig{
@@ -159,6 +160,14 @@ func run(
 		defer output_printers.PrintEnclaveId(enclaveCtx.GetEnclaveID())
 	}
 
+	if strings.HasPrefix(startosisScriptOrModulePath, githubDomainPrefix) {
+		err = executeRemoteModule(enclaveCtx, startosisScriptOrModulePath, serializedJsonArgs, dryRun)
+		if err != nil {
+			return stacktrace.Propagate(err, "An error occurred while running the module '%v'", startosisScriptOrModulePath)
+		}
+		return nil
+	}
+
 	fileOrDir, err := os.Stat(startosisScriptOrModulePath)
 	if err != nil {
 		return stacktrace.Propagate(err, "There was an error reading file or module from disk at '%v'", startosisScriptOrModulePath)
@@ -197,6 +206,11 @@ func validateScriptOrModulePath(_ context.Context, _ *flags.ParsedFlags, args *a
 	scriptOrModulePath = strings.TrimSpace(scriptOrModulePath)
 	if scriptOrModulePath == "" {
 		return stacktrace.NewError("Received an empty '%v'. It should be a non empty string.", scriptOrModulePathKey)
+	}
+
+	if strings.HasPrefix(scriptOrModulePath, githubDomainPrefix) {
+		// if it's a Github path we don't validate further, the APIC will do it for us
+		return nil
 	}
 
 	fileInfo, err := os.Stat(scriptOrModulePath)
@@ -250,6 +264,20 @@ func executeModule(enclaveCtx *enclaves.EnclaveContext, modulePath string, seria
 	return nil
 }
 
+func executeRemoteModule(enclaveCtx *enclaves.EnclaveContext, moduleId string, serializedParams string, dryRun bool) error {
+	executionResponse, err := enclaveCtx.ExecuteStartosisRemoteModule(moduleId, serializedParams, dryRun)
+	if err != nil {
+		return stacktrace.Propagate(err, "An unexpected error occurred executing the Startosis module '%s'", moduleId)
+	}
+
+	err = validateExecutionResponse(executionResponse, moduleId, moduleArgForLogging, dryRun)
+	if err != nil {
+		return stacktrace.Propagate(err, "Ran into a few errors while interpreting, validating or executing the module '%v' with dry-run set to '%v'", moduleId, dryRun)
+	}
+
+	return nil
+}
+
 func validateExecutionResponse(executionResponse *kurtosis_core_rpc_api_bindings.ExecuteStartosisResponse, scriptOrModulePath string, scriptOrModuleArg string, dryRun bool) error {
 	if executionResponse.GetInterpretationError() != kurtosisNoInterpretationError {
 		return stacktrace.NewError("There was an error interpreting the Startosis %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetInterpretationError().GetErrorMessage())
@@ -261,7 +289,7 @@ func validateExecutionResponse(executionResponse *kurtosis_core_rpc_api_bindings
 	var scriptOutputLines []string
 	concatenatedKurtosisInstructions := make([]string, len(executionResponse.GetKurtosisInstructions()))
 	for idx, instruction := range executionResponse.GetKurtosisInstructions() {
-		concatenatedKurtosisInstructions[idx] = instruction.GetExecutableInstruction()
+		concatenatedKurtosisInstructions[idx] = output_printers.FormatInstruction(instruction)
 		if instruction.InstructionResult != kurtosisInstructionNoResult {
 			scriptOutputLines = append(scriptOutputLines, instruction.GetInstructionResult())
 		}